Fix volume deserialization: Dhan returns volume as float64

+++ go/internal/run/worker.go
@@ -140,7 +140,7 @@ type candleResp struct {
 	High      []float64 `json:"high"`
 	Low       []float64 `json:"low"`
 	Close     []float64 `json:"close"`
-	Volume    []int64   `json:"volume"`
+	Volume    []float64 `json:"volume"`
 	Timestamp []int64   `json:"timestamp"`
 }
 
@@ -246,7 +246,7 @@ func (w *Worker) upsertCandles(secID, seg, interval string, c candleResp) error 
 	for i := range c.Close {
 		_, err := stmt.Exec(secID, seg, interval,
 			c.Timestamp[i],
-			c.Open[i], c.High[i], c.Low[i], c.Close[i], c.Volume[i])
+			c.Open[i], c.High[i], c.Low[i], c.Close[i], int64(c.Volume[i]))
 		if err != nil {
 			return err
 		}
